Remove unused MihailovaSystem declaration
The MihailovaSystem attribute is already declared for IsAutomatonGroup; and
only a method for that filter is installed. The removed declaration is for
IsSelfSimGroup, and no methods for it are ever installed.

This fixes a bug where the documentation entry for MihailovaSystem could be
empty, depending on which of the two declarations buildman.pe processed first.